improve swnlm_cuda, final version

+++ apps/swnlm_cuda/swnlmcuda.cpp
@@ -10,10 +10,10 @@ using namespace std;
 int main()
 {
     const string filename = "/home/bob/Documents/Uni/BachelorThesis/NonLocalMeans/images/mandril.tif";// "../../../images/mandril.tif";
-    const short sigma = 30;
+    const double sigma = 30;
     const int searchRadius = 10;
     const int neighborRadius = 3;
     const bool showImg = true;
 
-    testNLM(filename, sigma, searchRadius, neighborRadius, &swnlmcuda<short>, showImg);
+    testNLM(filename, sigma, searchRadius, neighborRadius, &swnlmcuda<double>, showImg);
 }
